feat: revert of the revert of "fix: swapper transition states (#7917)" (#8002)
