+++ src/state/slices/portfolioSlice/portfolioSlice.ts
@@ -32,9 +32,6 @@ export const portfolio = createSlice({
     clear: () => {
       return initialState
     },
-    setIsAccountsMetadataLoading: (state, { payload }: { payload: boolean }) => {
-      state.isAccountsMetadataLoading = payload
-    },
     setIsAccountMetadataLoading: (
       state,
       { payload }: { payload: { accountId: AccountId; isLoading: boolean } },
