+++ src/state/slices/portfolioSlice/portfolioSliceCommon.ts
@@ -54,7 +54,6 @@ export type ConnectWallet = {
 }
 
 export type Portfolio = {
-  isAccountsMetadataLoading: boolean
   isAccountMetadataLoadingByAccountId: Record<AccountId, boolean>
   /**
    * lookup of accountId -> accountMetadata
@@ -75,7 +74,6 @@ export type Portfolio = {
 }
 
 export const initialState: Portfolio = {
-  isAccountsMetadataLoading: false,
   isAccountMetadataLoadingByAccountId: {},
   accounts: {
     byId: {},
