+++ src/test/mocks/store.ts
@@ -46,7 +46,6 @@ export const mockStore: ReduxState = {
       version: 0,
       rehydrated: false,
     },
-    isAccountsMetadataLoading: false,
     isAccountMetadataLoadingByAccountId: {},
     accounts: {
       byId: {},
